Move conversions to rj_convert
