+++ rj_convert/include/rj_convert/testing/ros_convert_testing.hpp
@@ -1,12 +1,12 @@
 #include <gtest/gtest.h>
 
-#include "rj_common/ros_convert.hpp"
+#include "../ros_convert.hpp"
 
 #define TEST_LOSSLESS_CONVERT_ROS_VALUE(CppType, RosType, ros_value)           \
     {                                                                          \
         const auto& tmp_value = ros_value;                                     \
         EXPECT_EQ(                                                             \
-            (rj_common::convert_to_ros<CppType, RosType>(rj_common::convert_from_ros<CppType, RosType>(tmp_value))), \
+            (rj_convert::convert_to_ros<CppType, RosType>(rj_convert::convert_from_ros<CppType, RosType>(tmp_value))), \
             tmp_value);                                                        \
     }
 
@@ -14,6 +14,6 @@
     {                                                                          \
         const auto& tmp_value = cpp_value;                                     \
         EXPECT_EQ(                                                             \
-            (rj_common::convert_from_ros<CppType, RosType>(rj_common::convert_to_ros<CppType, RosType>(tmp_value))), \
+            (rj_convert::convert_from_ros<CppType, RosType>(rj_convert::convert_to_ros<CppType, RosType>(tmp_value))), \
             tmp_value);                                                        \
     }
