+++ rj_convert/testing/ros_convert_test.cpp
@@ -1,28 +1,30 @@
-#include "rj_common/ros_convert.hpp"
+#include <rj_convert/ros_convert.hpp>
 
 #include <gtest/gtest.h>
 
-#include "rj_common/testing/ros_convert_testing.hpp"
+#include <rj_convert/testing/ros_convert_testing.hpp>
+
+struct MockTime {
+    rclcpp::Time time;
+};
+
+namespace rj_convert {
+
+template<> struct RosConverter<MockTime, rclcpp::Time> {
+    static rclcpp::Time to_ros(const MockTime& from) {
+        return from.time;
+    }
+
+    static MockTime from_ros(const rclcpp::Time& from) {
+        return MockTime{from};
+    }
+};
+
+} // namespace rj_convert
 
 TEST(ROSConvert, primitives_have_lossless_convert) {
     TEST_LOSSLESS_CONVERT_ROS_VALUE(int, int, 5);
     TEST_LOSSLESS_CONVERT_CPP_VALUE(int, int, 5);
-}
-
-TEST(ROSConvert, time_lossless_convert) {
-    TEST_LOSSLESS_CONVERT_ROS_VALUE(RJ::Time, rclcpp::Time,
-                                    rclcpp::Time(123456));
-    TEST_LOSSLESS_CONVERT_CPP_VALUE(RJ::Time, rclcpp::Time, RJ::now());
-}
-
-TEST(ROSConvert, duration_lossless_convert) {
-    // We don't expect perfect equality for seconds, because float comparisons.
-    using Cvt = rj_common::RosConverter<RJ::Seconds, rclcpp::Duration>;
-    EXPECT_NEAR(Cvt::from_ros(Cvt::to_ros(RJ::Seconds(1.0))).count(), 1.0,
-                1e-6);
-    EXPECT_NEAR(
-        Cvt::to_ros(Cvt::from_ros(rclcpp::Duration(12345))).nanoseconds(),
-        12345, 1);
 }
 
 TEST(ROSConvert, vector_primitive) {
@@ -34,14 +36,14 @@ TEST(ROSConvert, vector_primitive) {
 TEST(ROSConvert, vector_ros_type) {
     std::vector<rclcpp::Time> vec_ros{rclcpp::Time(1), rclcpp::Time(2),
                                       rclcpp::Time(3)};
-    TEST_LOSSLESS_CONVERT_ROS_VALUE(std::vector<RJ::Time>,
+    TEST_LOSSLESS_CONVERT_ROS_VALUE(std::vector<MockTime>,
                                     std::vector<rclcpp::Time>, vec_ros);
 }
 
 TEST(ROSConvert, vector_nested) {
     std::vector<std::vector<rclcpp::Time>> vec_ros{
         {rclcpp::Time(1), rclcpp::Time(2)}, {rclcpp::Time(3)}};
-    TEST_LOSSLESS_CONVERT_ROS_VALUE(std::vector<std::vector<RJ::Time>>,
+    TEST_LOSSLESS_CONVERT_ROS_VALUE(std::vector<std::vector<MockTime>>,
                                     std::vector<std::vector<rclcpp::Time>>,
                                     vec_ros);
 }
